New chart project
Removed the charts from gui and added the chart project.  DO NOT COMPILE
ON LINUX YET

+++ GsTLAppli/gui/QWidget_value_accessors/qspin_box_accessor.cpp
@@ -169,7 +169,7 @@ bool QDoubleSpinBox_accessor::set_value( const std::string& str ) {
   
   QString qstr_value = elem.attribute( "value" );
 
-  int val = qstr_value.toInt();
+  double val = qstr_value.toDouble();
   spin_box_->setValue( val );
   return true;
 }
